NO JIRA. Replaced record separator for csv.
Replace the default x0d x0a record separator with just x0a, the Linux standard.
The mail program seems to replace x0a with x0d x0a, as this is the standard in
web protocols. This left us with an extra newline in the CSV output and therefore
with blank rows.